updated upload page design
