+++ sensitive-information-blurring-webapp/src/components/faq.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import NavBar from './navbar'
+
+export default function FAQ() {
+    return (
+        <>
+            <NavBar title="Frequently Asked Questions" />
+        </>
+    )
+}
