+++ sensitive-information-blurring-webapp/src/components/navbar.jsx
@@ -4,18 +4,18 @@ import Navbar from 'react-bootstrap/Navbar';
 import '../css/navbarstyles.css';
 
 
-function NavBar() {
+function NavBar(props) {
     return (
         <>
-            <Navbar expand="md" className='navbar'>
+            <Navbar collapseOnSelect expand="md" className='navbar' fixed='top'>
                 <Container>
-                    <Navbar.Brand href="#home" className='navTitle'>Sensitive Info Blurring</Navbar.Brand>
+                    <Navbar.Brand href="" className='navTitle'>{props.title}</Navbar.Brand>
                     <Navbar.Toggle aria-controls="basic-navbar-nav" />
-                    <Navbar.Collapse id="basic-navbar-nav">
+                    <Navbar.Collapse id="responsive-navbar-nav">
                         <Nav className="ms-auto">
-                            <Nav.Link href="#" className='navLink'>Home</Nav.Link>
-                            <Nav.Link href="#" className='navLink'>Privacy Policy</Nav.Link>
-                            <Nav.Link href="#" className='navLink'>FAQ</Nav.Link>
+                            <Nav.Link href="/" className='navLink'>Home</Nav.Link>
+                            <Nav.Link href="/privacy" className='navLink'>Privacy Policy</Nav.Link>
+                            <Nav.Link href="/faq" className='navLink'>FAQ</Nav.Link>
                         </Nav>
                     </Navbar.Collapse>
                 </Container>
