+++ sensitive-information-blurring-webapp/src/components/privacy.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import NavBar from './navbar'
+
+export default function Privacy() {
+  return (
+    <>
+        <NavBar title="Privacy Policy"/>
+    </>
+  )
+}
