+++ sensitive-information-blurring-webapp/src/css/navbarstyles.css
@@ -1,10 +1,4 @@
-.navbar {
-    width: 100%;
-    min-width: 475px;
-}
-
 .navTitle {
-    font-size: 2rem;
     font-weight: 600;
 }
