+++ sensitive-information-blurring-webapp/src/index.css
